Initial restructure and renaming of subject locations controller to subject controller
