+++ helm_deploy/hmpps-electronic-monitoring-crime-matching-ui/files/stubs/__files/crime-matching-api-get-subjects-query-foo-page1-200.json
@@ -1,6 +1,7 @@
 {
   "data": [
     {
+      "personId": "1",
       "nomisId": "Nomis 1",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -12,6 +13,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "2",
       "nomisId": "Nomis 2",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -23,6 +25,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "3",
       "nomisId": "Nomis 3",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -34,6 +37,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "4",
       "nomisId": "Nomis 4",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -45,6 +49,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "5",
       "nomisId": "Nomis 5",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -56,6 +61,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "6",
       "nomisId": "Nomis 6",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -67,6 +73,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "7",
       "nomisId": "Nomis 7",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -78,6 +85,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "8",
       "nomisId": "Nomis 8",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -89,6 +97,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "9",
       "nomisId": "Nomis 9",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
@@ -100,6 +109,7 @@
       "tagPeriodEndDate": null
     },
     {
+      "personId": "10",
       "nomisId": "Nomis 10",
       "name": "John",
       "dateOfBirth": "2000-12-01T00:00:00.000Z",
