+++ integration_tests/e2e/locationData/subjects.page.cy.ts
@@ -100,25 +100,25 @@ context('Location Data', () => {
           '',
           'Nomis 1',
           'John',
-          '2000-12-01T00:00:00.000Z',
+          '01/12/2000 00:00',
           '123 Street',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
           '',
           '123456',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
           '',
         ],
         [
           '',
           'Nomis 2',
           'Lee',
-          '2000-12-01T00:00:00.000Z',
+          '01/12/2000 00:00',
           '456 Avenue',
-          '2024-12-01T00:00:00.000Z',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
+          '01/12/2024 00:00',
           '654321',
-          '2024-12-01T00:00:00.000Z',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
+          '01/12/2024 00:00',
         ],
       ])
       page.dataTable.shouldNotHavePagination()
@@ -134,6 +134,7 @@ context('Location Data', () => {
         response: {
           data: [
             {
+              personId: '1',
               nomisId: 'Nomis 1',
               name: 'John',
               dateOfBirth: '2000-12-01T00:00:00.000Z',
@@ -157,6 +158,7 @@ context('Location Data', () => {
         response: {
           data: [
             {
+              personId: '2',
               nomisId: 'Nomis 2',
               name: 'Lee',
               dateOfBirth: '2000-12-01T00:00:00.000Z',
@@ -202,12 +204,12 @@ context('Location Data', () => {
           '',
           'Nomis 1',
           'John',
-          '2000-12-01T00:00:00.000Z',
+          '01/12/2000 00:00',
           '123 Street',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
           '',
           '123456',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
           '',
         ],
       ])
@@ -228,13 +230,13 @@ context('Location Data', () => {
           '',
           'Nomis 2',
           'Lee',
-          '2000-12-01T00:00:00.000Z',
+          '01/12/2000 00:00',
           '456 Avenue',
-          '2024-12-01T00:00:00.000Z',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
+          '01/12/2024 00:00',
           '654321',
-          '2024-12-01T00:00:00.000Z',
-          '2024-12-01T00:00:00.000Z',
+          '01/12/2024 00:00',
+          '01/12/2024 00:00',
         ],
       ])
       page.dataTable.shouldHavePagination()
