+++ integration_tests/index.d.ts
@@ -1,5 +1,5 @@
 import { StubCreateCrimeBatchQueryOptions, StubGetCrimeBatchesOptions } from './mockApis/crimeMatching/crimeBatches'
-import { StubCreateSubjectsQueryOptions, StubGetSubjectsQueryOptions } from './mockApis/crimeMatching/subjects'
+import { StubCreateSubjectsQueryOptions, StubGetSubjectsQueryOptions } from './mockApis/locationData/subjects'
 import { StubCreateSubjectLocationQueryOptions } from './mockApis/locationData/subjectLocations'
 
 declare global {
