+++ integration_tests/mockApis/locationData/subjects.ts
@@ -22,20 +22,20 @@ type StubCreateSubjectsQuery500Options = {
   response: string
 }
 
-type StubCreateSubjectQueryOptions =
+type StubCreateSubjectsQueryOptions =
   | StubCreateSubjectsQuery200Options
   | StubCreateSubjectsQuery400Options
   | StubCreateSubjectsQuery500Options
 
 // Default options returns a successful response with a mock queryExecutionId
-const defaultCreateSubjectQueryOptions: StubCreateSubjectQueryOptions = {
+const defaultCreateSubjectsQueryOptions: StubCreateSubjectsQueryOptions = {
   status: 200,
   response: {
     queryExecutionId: '1234',
   },
 }
 
-const stubCreateSubjectsQuery = (options: StubCreateSubjectQueryOptions = defaultCreateSubjectQueryOptions) =>
+const stubCreateSubjectsQuery = (options: StubCreateSubjectsQueryOptions = defaultCreateSubjectsQueryOptions) =>
   stubFor({
     request: {
       method: 'POST',
@@ -53,18 +53,23 @@ const stubCreateSubjectsQuery = (options: StubCreateSubjectQueryOptions = defaul
 type StubGetSubjectsQuery200Options = {
   status: 200
   query: string
-  response: Array<{
-    personId: string
-    nomisId: string
-    name: string
-    dateOfBirth: string
-    address: string
-    orderStartDate: string
-    orderEndDate: string
-    deviceId: string
-    tagPeriodStartDate: string
-    tagPeriodEndDate: string
-  }>
+  response: {
+    data: Array<{
+      personId: string
+      nomisId: string
+      name: string
+      dateOfBirth: string
+      address: string
+      orderStartDate: string
+      orderEndDate: string
+      deviceId: string
+      tagPeriodStartDate: string
+      tagPeriodEndDate: string
+    }>
+    pageCount: number
+    pageNumber: number
+    pageSize: number
+  }
 }
 
 type StubGetSubjectsQuery404Options = {
@@ -75,13 +80,18 @@ type StubGetSubjectsQuery404Options = {
 
 type StubGetSubjectsQueryOptions = StubGetSubjectsQuery200Options | StubGetSubjectsQuery404Options
 
-const defaultGetSubjectQueryOptions: StubGetSubjectsQueryOptions = {
+const defaultGetSubjectsQueryOptions: StubGetSubjectsQueryOptions = {
   status: 200,
   query: '.*',
-  response: [],
+  response: {
+    data: [],
+    pageCount: 1,
+    pageNumber: 1,
+    pageSize: 10,
+  },
 }
 
-const stubGetSubjectsQuery = (options: StubGetSubjectsQueryOptions = defaultGetSubjectQueryOptions) =>
+const stubGetSubjectsQuery = (options: StubGetSubjectsQueryOptions = defaultGetSubjectsQueryOptions) =>
   stubFor({
     request: {
       method: 'GET',
@@ -96,4 +106,4 @@ const stubGetSubjectsQuery = (options: StubGetSubjectsQueryOptions = defaultGetS
     },
   })
 
-export { stubCreateSubjectsQuery, stubGetSubjectsQuery, StubCreateSubjectQueryOptions, StubGetSubjectsQueryOptions }
+export { stubCreateSubjectsQuery, stubGetSubjectsQuery, StubCreateSubjectsQueryOptions, StubGetSubjectsQueryOptions }
