+++ server/types/locationData/subjectLocations.ts
@@ -1,0 +1,20 @@
+type CreateSubjectLocationsQueryRequestDto = {
+  personId: string
+  fromDate: DateAndTimeInput
+  toDate: DateAndTimeInput
+  orderStartDate: string
+  orderEndDate: string
+}
+
+type CreateSubjectLocationsQueryResponseDto = {
+  queryExecutionId: string
+}
+
+type DateAndTimeInput = {
+  date: string
+  hour: string
+  minute: string
+  second: string
+}
+
+export { CreateSubjectLocationsQueryRequestDto, CreateSubjectLocationsQueryResponseDto }
