+++ server/views/pages/locationData/index.njk
@@ -98,7 +98,7 @@
       </form>
     </div>
 
-    <form action="/location-data/subjects/location-search" method="post" novalidate id="subject-location-form">
+    <form action="/location-data/subject/location-search" method="post" novalidate id="subject-location-form">
       <div class="govuk-grid-column-one-third">
         <h2>Location Data</h2>
           <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
